+++ src/main/java/net/geforcemods/securitycraft/blocks/ProjectorBlock.java
@@ -2,7 +2,6 @@ package net.geforcemods.securitycraft.blocks;
 
 import java.util.List;
 import java.util.Random;
-import java.util.stream.Stream;
 
 import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.blockentities.ProjectorBlockEntity;
@@ -17,6 +16,7 @@ import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.inventory.container.INamedContainerProvider;
 import net.minecraft.item.BlockItemUseContext;
 import net.minecraft.item.ItemStack;
+import net.minecraft.state.BooleanProperty;
 import net.minecraft.state.DirectionProperty;
 import net.minecraft.state.StateContainer.Builder;
 import net.minecraft.state.properties.BlockStateProperties;
@@ -28,7 +28,6 @@ import net.minecraft.util.Mirror;
 import net.minecraft.util.Rotation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.util.math.shapes.IBooleanFunction;
 import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.util.math.shapes.VoxelShapes;
@@ -45,14 +44,19 @@ import net.minecraftforge.fml.network.NetworkHooks;
 public class ProjectorBlock extends DisguisableBlock {
 	private static final IFormattableTextComponent TOOLTIP = new TranslationTextComponent("tooltip.securitycraft:projector").setStyle(Utils.GRAY_STYLE);
 	public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
-	private static final VoxelShape NORTH = Stream.of(Block.box(12, 0, 12, 15, 2, 15), Block.box(1, 2, 1, 15, 7, 15), Block.box(9, 2, 15, 14, 7, 16), Block.box(12, 0, 1, 15, 2, 4), Block.box(1, 0, 1, 4, 2, 4), Block.box(1, 0, 12, 4, 2, 15)).reduce((v1, v2) -> VoxelShapes.join(v1, v2, IBooleanFunction.OR)).orElse(VoxelShapes.block());
-	private static final VoxelShape SOUTH = Stream.of(Block.box(1, 0, 1, 4, 2, 4), Block.box(1, 2, 1, 15, 7, 15), Block.box(2, 2, 0, 7, 7, 1), Block.box(1, 0, 12, 4, 2, 15), Block.box(12, 0, 12, 15, 2, 15), Block.box(12, 0, 1, 15, 2, 4)).reduce((v1, v2) -> VoxelShapes.join(v1, v2, IBooleanFunction.OR)).orElse(VoxelShapes.block());
-	private static final VoxelShape WEST = Stream.of(Block.box(12, 0, 1, 15, 2, 4), Block.box(1, 2, 1, 15, 7, 15), Block.box(15, 2, 2, 16, 7, 7), Block.box(1, 0, 1, 4, 2, 4), Block.box(1, 0, 12, 4, 2, 15), Block.box(12, 0, 12, 15, 2, 15)).reduce((v1, v2) -> VoxelShapes.join(v1, v2, IBooleanFunction.OR)).orElse(VoxelShapes.block());
-	private static final VoxelShape EAST = Stream.of(Block.box(1, 0, 12, 4, 2, 15), Block.box(1, 2, 1, 15, 7, 15), Block.box(0, 2, 9, 1, 7, 14), Block.box(12, 0, 12, 15, 2, 15), Block.box(12, 0, 1, 15, 2, 4), Block.box(1, 0, 1, 4, 2, 4)).reduce((v1, v2) -> VoxelShapes.join(v1, v2, IBooleanFunction.OR)).orElse(VoxelShapes.block());
+	public static final BooleanProperty HANGING = BlockStateProperties.HANGING;
+	private static final VoxelShape FLOOR_NORTH = VoxelShapes.or(Block.box(12, 0, 12, 15, 2, 15), Block.box(1, 2, 1, 15, 7, 15), Block.box(9, 2, 15, 14, 7, 16), Block.box(12, 0, 1, 15, 2, 4), Block.box(1, 0, 1, 4, 2, 4), Block.box(1, 0, 12, 4, 2, 15));
+	private static final VoxelShape FLOOR_SOUTH = VoxelShapes.or(Block.box(1, 0, 1, 4, 2, 4), Block.box(1, 2, 1, 15, 7, 15), Block.box(2, 2, 0, 7, 7, 1), Block.box(1, 0, 12, 4, 2, 15), Block.box(12, 0, 12, 15, 2, 15), Block.box(12, 0, 1, 15, 2, 4));
+	private static final VoxelShape FLOOR_WEST = VoxelShapes.or(Block.box(12, 0, 1, 15, 2, 4), Block.box(1, 2, 1, 15, 7, 15), Block.box(15, 2, 2, 16, 7, 7), Block.box(1, 0, 1, 4, 2, 4), Block.box(1, 0, 12, 4, 2, 15), Block.box(12, 0, 12, 15, 2, 15));
+	private static final VoxelShape FLOOR_EAST = VoxelShapes.or(Block.box(1, 0, 12, 4, 2, 15), Block.box(1, 2, 1, 15, 7, 15), Block.box(0, 2, 9, 1, 7, 14), Block.box(12, 0, 12, 15, 2, 15), Block.box(12, 0, 1, 15, 2, 4), Block.box(1, 0, 1, 4, 2, 4));
+	private static final VoxelShape CEILING_NORTH = VoxelShapes.or(Block.box(1, 6, 1, 15, 11, 15), Block.box(9, 6, 15, 14, 11, 16), Block.box(7, 11, 4, 9, 16, 6));
+	private static final VoxelShape CEILING_SOUTH = VoxelShapes.or(Block.box(1, 6, 1, 15, 11, 15), Block.box(2, 6, 0, 7, 11, 1), Block.box(7, 11, 10, 9, 16, 12));
+	private static final VoxelShape CEILING_WEST = VoxelShapes.or(Block.box(1, 6, 1, 15, 11, 15), Block.box(15, 6, 2, 16, 11, 7), Block.box(4, 11, 7, 6, 16, 9));
+	private static final VoxelShape CEILING_EAST = VoxelShapes.or(Block.box(1, 6, 1, 15, 11, 15), Block.box(0, 6, 9, 1, 11, 14), Block.box(10, 11, 7, 12, 16, 9));
 
 	public ProjectorBlock(Block.Properties properties) {
 		super(properties);
-		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH));
+		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH).setValue(HANGING, false));
 	}
 
 	@Override
@@ -61,16 +65,30 @@ public class ProjectorBlock extends DisguisableBlock {
 
 		if (disguisedState.getBlock() != this)
 			return disguisedState.getShape(world, pos, ctx);
+		else if (!disguisedState.getValue(HANGING)) {
+			switch (disguisedState.getValue(FACING)) {
+				case NORTH:
+					return FLOOR_NORTH;
+				case EAST:
+					return FLOOR_EAST;
+				case SOUTH:
+					return FLOOR_SOUTH;
+				case WEST:
+					return FLOOR_WEST;
+				default:
+					return VoxelShapes.block();
+			}
+		}
 		else {
 			switch (disguisedState.getValue(FACING)) {
 				case NORTH:
-					return NORTH;
+					return CEILING_NORTH;
 				case EAST:
-					return EAST;
+					return CEILING_EAST;
 				case SOUTH:
-					return SOUTH;
+					return CEILING_SOUTH;
 				case WEST:
-					return WEST;
+					return CEILING_WEST;
 				default:
 					return VoxelShapes.block();
 			}
@@ -141,13 +159,13 @@ public class ProjectorBlock extends DisguisableBlock {
 		return getStateForPlacement(ctx.getLevel(), ctx.getClickedPos(), ctx.getClickedFace(), ctx.getClickLocation().x, ctx.getClickLocation().y, ctx.getClickLocation().z, ctx.getPlayer());
 	}
 
-	public BlockState getStateForPlacement(World world, BlockPos pos, Direction facing, double hitX, double hitY, double hitZ, PlayerEntity placer) {
-		return defaultBlockState().setValue(FACING, placer.getDirection().getOpposite());
+	public BlockState getStateForPlacement(World world, BlockPos pos, Direction clickedFace, double hitX, double hitY, double hitZ, PlayerEntity placer) {
+		return defaultBlockState().setValue(FACING, placer.getDirection().getOpposite()).setValue(HANGING, clickedFace == Direction.DOWN);
 	}
 
 	@Override
 	protected void createBlockStateDefinition(Builder<Block, BlockState> builder) {
-		builder.add(FACING);
+		builder.add(FACING, HANGING);
 	}
 
 	@Override