+++ Changelog.md
@@ -7,6 +7,7 @@
 - New: The Allowlist and Denylist Module can now contain teams, meaning every player on a team that is on the list will be allowed/denied
 - New: Modules can now be toggled off or on without removing them from the block, by clicking the respective button in the Universal Block Modifier screen
 - New: Reinforced End Rod
+- New: The Projector can now be placed on the ceiling, allowing for downwards projection
 - Change: Codebreakers will now always open briefcases when in Creative mode without a chance of failing
 - Change: The Sentry's body is now solid if the sentry is not disguised
 - Change: The Fake Lava/Water Buckets are now created in a brewing stand, by putting the potion in the top slot, and the bucket(s) in the bottom slot(s)